+++ qiskit/wrapper/__init__.py
@@ -17,4 +17,5 @@ refer to the documentation of each component and use them separately.
 
 from ._wrapper import (available_backends, local_backends, remote_backends,
                        get_backend, compile, execute, register, unregister,
-                       registered_providers, load_qasm_string, load_qasm_file)
+                       registered_providers, load_qasm_string, load_qasm_file,
+                       least_busy)